+++ Changelog.md
@@ -1,6 +1,7 @@
 # Change log
 
 ### vNext
+- fix shallowEqual bug.
 
 ### 1.4.11
 - Replace string refs with callback refs [#908](https://github.com/apollographql/react-apollo/pull/908)