+++ Changelog.md
@@ -1,6 +1,7 @@
 # Change log
 
 ### vNext
+- fix shallowEqual bug.
 - Added notifyOnNetworkStatusChange to QueryOpts and MutationOpts Typesccript definitions [#1034](https://github.com/apollographql/react-apollo/pull/1034)
 
 ### 1.4.15