+++ Changelog.md
@@ -1,6 +1,8 @@
 # Change log
 
 ### vNext
+
+### 1.4.16
 - upgrade to react-16 
 - fix shallowEqual bug.
 - Added notifyOnNetworkStatusChange to QueryOpts and MutationOpts Typesccript definitions [#1034](https://github.com/apollographql/react-apollo/pull/1034)